added refacto and statics links
