+++ client/src/assets/SideNavBarLinks/AdminLinks.tsx
@@ -1,10 +1,10 @@
 import { ReactElement } from "react";
-import { AccountCircleOutlined } from "@mui/icons-material";
+import AccountCircleOutlinedIcon from "@mui/icons-material/AccountCircleOutlined";
 
 export const adminLinks: { name: string; url: string; icon: ReactElement }[] = [
   {
     name: "Utilisateurs",
     url: "utilisateurs",
-    icon: <AccountCircleOutlined />,
+    icon: <AccountCircleOutlinedIcon />,
   },
 ];
