+++ client/src/assets/SideNavBarLinks/StaticLinks.tsx
@@ -1,0 +1,11 @@
+import { ReactElement } from "react";
+import LogoutOutlinedIcon from "@mui/icons-material/LogoutOutlined";
+
+export const staticLinks: { name: string; url: string; icon: ReactElement }[] =
+  [
+    {
+      name: "Déconnexion",
+      url: "/deconnexion",
+      icon: <LogoutOutlinedIcon />,
+    },
+  ];
