+++ client/src/types/SideNavBarTypes.ts
@@ -1,0 +1,12 @@
+import { ReactElement } from "react";
+
+export type SideNavBarPropsType = {
+  teamLinks: linkType[];
+  teamDashboardUrl: string;
+};
+
+export type linkType = {
+  name: string;
+  url: string;
+  icon: ReactElement;
+};
